support desimal

+++ sales/models.py
@@ -27,7 +27,7 @@ class Sale(BaseModel):
 class SaleItem(BaseModel):
     sale = models.ForeignKey(Sale, on_delete=models.CASCADE, related_name='items')
     product = models.ForeignKey(Product, on_delete=models.PROTECT)
-    qty = models.PositiveIntegerField()
+    qty = models.DecimalField(max_digits=12, decimal_places=3)
     unit_price = models.DecimalField(max_digits=12, decimal_places=2)
     line_total = models.DecimalField(max_digits=14, decimal_places=2)
